# lib/swagger/doc/internal.yaml
swagger: '2.0'
info:
  description: Helo Whatsapp API documentation
  version: 1.0.0
  title: Helo Whatsapp
  termsOfService: ''
  contact:
    name: vivaconnect support
    email: support@vivaconnect.co
...
schemes:
  - http
paths:
...
  /business/profile:
    get:
      description: Gets business profile data
      summary: Get business profile
      tags:
        - WABA
      operationId: getbusinessprofile
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
      responses:
        '200':
          description: business profie data
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 2000
              msg: Success
              data:
                wabaInformationId: f742d535-4161-4a11-df98-0c97154b72sde
                phoneNumber: '1234567890'
                phoneCode: '91'
                facebookManagerId: my-business manager-id
                businessName: vivaconnect
                email: 123@gmail.com
                businessCategory: Other
                state: maharashtra
                whatsappStatus: available up-date
                description: available 24/7
                profilePhotoUrl: www.google.com/image
                address: some address
                country: IN
                wabaProfileSetupStatus: Profile Incomplete / Pending For Submission
                businessManagerVerified: true
                phoneVerified: true
                city: mumbai
                postalCode: '400009'
                serviceProviderId: f1d44200-4b9d-4901-ae49-5035e0b14a5d
                serviceProviderUserAccountId: asfwevawvw33423sdfsdrf32
                serviceProviderName: Tyntec
                apiKey: test123
                webhookPostUrl: 'http://localhost:3000/helowhatsapp/message'
                optinText: hello-viva
                businessProfileCompletionStatus: true
                businessAccessProfileCompletionStatus: true
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
    post:
      description: API to add or update business profile
      summary: add or update business profile
      tags:
        - WABA
      operationId: add/updatebusinessprofile
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/add~1updatebusinessprofilerequest'
      responses:
        '200':
          description: Returns Status of business profile info completion
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 2000
              msg: Success
              data:
                businessProfileCompletionStatus: true
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
  /business/profile/markManagerVerified:
    post:
      summary: mark manager verified
      description: This API marks manager verified
      tags:
        - WABA
      operationId: markmanagerverified
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/markmanagerverifiedrequest'
      responses:
        '200':
          description: Returns Status of business verification completion
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 2000
              msg: Success
              data:
                businessVerificationCompletionStatus: true
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 4001
              msg: Unauthorized access.
              data: null
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
  /business/profile/serviceProvider:
    put:
      description: API is used for updating service provider id
      summary: update service provier id
      tags:
        - WABA
      operationId: updateServiceProviderId
      deprecated: false
      produces:
        - application/json
      parameters:
        - name: Authorization
          in: header
          required: true
          type: string
          description: user can obtain auth token by using login API
        - name: Body
          in: body
          required: true
          description: ''
          schema:
            $ref: '#/definitions/updateServiceProviderIdRequest'
      responses:
        '200':
          description: Returns Service Provider Id
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
            example:
              code: 2000
              msg: Success
              data:
                serviceProviderId: e76a602e-37e3-4c5d-898f-56bf0c880f93
        '400':
          description: Returns invalid request error in error array
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: array
            example:
              code: 4000
              msg: Invalid request
              data: null
              error:
                - invalid body error here
        '401':
          description: Invalid token in header
          headers: {}
          schema:
            type: string
            example: Unauthorized
        '500':
          description: in case of some exceptional error
          headers: {}
          schema:
            properties:
              code:
                type: number
              msg:
                type: string
              data:
                type: object
              error:
                type: object
            example:
              code: 5000
              msg: Something went wrong. Please try again later.
              data: {}
              error: {}
      security: []
...
definitions:
...
  markmanagerverifiedrequest:
    title: markmanagerverifiedrequest
    example:
      businessManagerVerified: true
    type: object
    properties:
      businessManagerVerified:
        type: boolean
    required:
      - businessManagerVerified
  updateServiceProviderIdRequest:
    title: updateServiceProviderIdRequest
    example:
      serviceProviderId: e76a602e-37e3-4c5d-898f-56bf0c880f93
    type: object
    properties: {}
...